FOGT-5: opslaan van de repositories. + In gamecreator het slepen van muren tegen elkaar heeft nu collission detection

+++ FirstOpenGLGameTest/src/be/davidcorp/domain/sprite/item/weapon/Weapon.java
@@ -28,7 +28,7 @@ public abstract class Weapon extends Item{
 		}
 	}
 	
-	public void updateSprite(int secondsMovedInGame){
+	public void updateSprite(float secondsMovedInGame){
 		super.updateSprite(secondsMovedInGame);
 		if(!attackCooldowner.isCoolDowned()){
 			attackCooldowner.progressCooldown();
